
\name{IMDb}
\alias{IMDb}
\docType{data}
\encoding{UTF-8}
\title{IMDb data}
\description{Data collected from IMDb.com on the ratings of 105 popular comedy movies.
}
\usage{data(IMDb)}
\format{ 
  A list of 2 objects: Y and name, where Y represents the data array of ratings and name represents the comedy movie names. Y is the of dimensionality 2 x 4 x 105 with ratings on 105 movies from female and male by age groups 0-18, 18-29, 30-44, 45+. 
}
\details{The data are publicly available on www.IMDb.com.}
\examples{

data(IMDb)

}

\keyword{datasets}
