\name{MatTransMix-package}
\alias{MatTransMix-package}
\docType{package}
\title{
Finite mixture modeling and model-based clustering of matrices based on matrix Gaussian mixture and matrix transformation mixture models.
}
\description{
The utility of this package is the clustering of random matrices. Finite mixture modeling and model-based clustering based on matrix Gaussian mixtures and matrix transformation mixtures are employed. 
}
\details{
\tabular{ll}{
Package: \tab MatTransMix \cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2017-02-09\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab no\cr
}

Function 'MatTrans.init' runs the initialization for the EM algorithm.

Function 'MatTrans.EM' runs the EM algorithm for matrix-variate mixtures to cluster matrices.

}
\author{ Xuwen Zhu and Volodymyr Melnykov.

Maintainer: Xuwen Zhu <xuwen.zhu@louisville.edu>
}

