
\name{Salary}
\alias{Salary}
\docType{data}
\encoding{UTF-8}
\title{Salary data}
\description{Data collected from the Chronicle of Higher Education web site reporting the average faculty salaries from 696 universities presented in the form of a 2 by 3 by 13 -dimensional tensor.}
\usage{data(Salary)}
\format{ 
  A list of 2 objects: Y, uni_info. Y represents the salary data array from 696 universities. uni_info has the university names, state and category. Y is of dimensionality 2 by 3 by 13 categorized by the following factors: gender (Male, Female), professor rank (Assistant, Associate, Full), and academic year (2003-2004,2015-2016). 
  }
\details{The data have been made publicly available by the Chronicle of Higher Education web site.}
\examples{

data(Salary)

}

\keyword{datasets}
