\name{print.object}
\alias{print.EM}
\alias{summary.EM}
\title{Functions for Printing or Summarizing Objects}
\description{
  \code{EM} classes for printing and summarizing objects.
}
\usage{
\method{print}{EM}(x, ...)
\method{summary}{EM}(object, ...)
}
\arguments{
  \item{x}{an object with the 'EM' class attributes.}
  \item{object}{an object with the 'EM' class attributes.}
  \item{\dots}{other possible options.}
}
\details{
  Some useful functions for printing and summarizing results.
}
\value{No return value, called for side effects}
\seealso{
  \code{MatTrans.EM}.
}

